[Fix] fix deadlock above on Android N cause by suspend all thread.
Above Android N, jit thread or other thread may modify the code protect property, when we modify it at the same time, conflict may occur. so we just suspend all thread to avoid this.

But, when we suspend all thread, if ourselves is suspend, the world will never restore. That is to say, when we stop the world, we must ensure every code is safe; So, this means that we can not call any jni method : translation from JNI and vm may cause thread suspend.

In a world, we must implement the acivate code with native code, and do not call any jni method bewteen suspend and resume.